pkg/storage: Add instrumented badgerdb package
Adds `database/badger` which provides an instrumented wrapper around a badger database. Only supports
a small subset of functionality so far that I'll keep adding to as and when I need it. So far handles
`View`, `Update`, `Get`, `Set` and `Value`. Transaction and item details are added to the spans.