pkg/server: Add bootstrapping for a gRPC server (#24)
Adds `server.ServeGRPC` which starts a new gRPC server listening on port 5000 that has
health checks & instrumentation for tracing, metrics, panic recovery, request validation
and context tag extration.

Creates a `GRPCService` abstraction for types that implement a gRPC service interface. They must
include a `Register` method that registers the type to the `grpc.Server` instance. This keeps the
specifics of the gRPC code close to the implementation.