pkg/closers: Add CloseFunc (#39)
Adds a `CloseFunc` method that returns an `io.Closer` implementation that calls a desired function. This can be
used in scenarios where you have a type that requires some form of closing but does not implement `io.Closer`
directly. For example, the elasticsearch `BulkIndexer` has a `Close(context.Context)` method.