pkg/server: Don't use package-level variable for default options
This fixes a bug causing tracing for gRPC to not work correctly, as when the `var` is instantiated, `tracing.New` has not
been called, so the global tracer is unset. This makes sure that we always get the options as late as possible so the global
tracer is set correctly.